i added juggle combos and you can't stop me
we can remove it if we decide the game would be better without it but i think it would be neat. it took like 2 additional lines of code. i actually had to remove one line to make it work.

oh also added a placeholder hitstun animation. well it rotates the whole object so it is not ideal.