begun on high and low blocking
implemented a very basic high and low system. right now it is based on an attack property rather than where on the hurtbox was hit. blocking is also working based on direction but right now it makes you completely invincible. have not added blockstun or knockback or block breaking yet.